bug fixes, code refactor, new features
1. Lightglue-based tracker now detects and matches keypoints inside an ROI instead of the whole frame, this improves tracking robustness.
2. Rewrite the fretboard detection class for better performance and readability.
3. Refactor part of the gui code and add some undocumented features.
4. Fix a bug related to multi-threading, which previously caused tracking to break when displaying frame at the same time.

+++ play2tab/video_processing/utils/visualize.py
@@ -70,5 +70,5 @@ def dashed_vertical_line(img, x, y0, y1, color, stroke=5, gap=5):
 def draw_houghlineP_batch(cdst, linesP):
     if linesP is not None:
         for i in range(0, len(linesP)):
-            l = linesP[i][0]
+            l = linesP[i][0].astype(int)
             cv2.line(cdst, (l[0], l[1]), (l[2], l[3]), (0,255,0), 1, cv2.LINE_AA)
